requirements.txt
numpy==1.21.6
pandas==1.3.5
alpaca-py==0.13.4
jupyter
apscheduler
python-dotenv